Fix a serious lag when moving point over a large amount of characters.
When scrolling, the values returned by window-start and window-end are
stale, and the point doesn't necessarily lie between the two. It used
to result in relative-line-numbers traversing *the whole file* in extreme
cases, creating an overlay for *every single line*.

If we number the lines only when window-start <= point <= window-end,
the problem goes away. As a bonus, we can get rid of the scrollbar hacks,
as scrolling is never slow enough to warrant them. Scrolling with the
scrollbar may still lag a bit, but it's a font-lock-mode problem.